+++ lib/src/test/utils_test.dart
@@ -2,29 +2,16 @@ import 'package:flutter/material.dart';
 
 import '../../flutter_modular.dart';
 
-class TestModule extends ChildModule {
-  final List<Bind> changeBinds;
-  final List<Router> changeRouters;
-
-  TestModule(this.changeBinds, this.changeRouters);
-
-  @override
-  List<Bind> get binds => changeBinds;
-
-  @override
-  List<Router> get routers => changeRouters;
-}
-
 void initModule(ChildModule module, {List<Bind> changeBinds}) {
-  // ChildModule changedModule = TestModule(changeBinds, module.routers);
-
-  for (var item in module.binds ?? []) {
+  final list = module.binds;
+  for (var item in list ?? []) {
     var dep = (changeBinds ?? []).firstWhere((dep) {
-      return item.inject.runtimeType == dep.inject.runtimeType;
+      return item.runtimeType == dep.runtimeType;
     }, orElse: () => null);
     if (dep != null) {
-      module.binds.remove(dep);
-      module.binds.add(item);
+      list.remove(item);
+      list.add(dep);
+      module.changeBinds(list);
     }
   }