+++ test/app/app_module.dart
@@ -6,11 +6,14 @@ import 'app_widget.dart';
 import 'guard/guard.dart';
 import 'modules/forbidden/forbidden_widget.dart';
 import 'modules/home/home_module.dart';
+import 'shared/ILocalRepository.dart';
+import 'shared/local_storage_shared.dart';
 
 class AppModule extends MainModule {
   @override
   List<Bind> get binds => [
         Bind((i) => AppBloc()),
+        Bind<ILocalStorage>((i) => LocalStorageSharePreference()),
       ];
 
   @override
@@ -26,7 +29,8 @@ class AppModule extends MainModule {
           module: HomeModule(),
           transition: TransitionType.fadeIn,
         ),
-        Router("/home", module: HomeModule(), guards: [MyGuard()]),
+        Router("/home", module: HomeModule()),
+        Router("/homeTwo", module: HomeModule(), guards: [MyGuard()]),
       ];
 
   @override