+++ test/app/shared/local_storage_shared.dart
@@ -1,0 +1,12 @@
+import 'ILocalRepository.dart';
+
+class LocalStorageSharePreference implements ILocalStorage {
+  @override
+  Future delete(String key) {}
+
+  @override
+  Future get(String key) {}
+
+  @override
+  Future put(String key, String value) {}
+}