+++ test/modular_inject_test.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:flutter_modular/flutter_modular.dart';
@@ -8,6 +7,8 @@ import 'app/app_bloc.dart';
 import 'app/app_module.dart';
 import 'app/modules/home/home_bloc.dart';
 import 'app/modules/home/home_module.dart';
+import 'app/shared/ILocalRepository.dart';
+import 'app/shared/local_storage_shared.dart';
 
 void main() {
   setUpAll(() {
@@ -24,16 +25,16 @@ void main() {
     });
 
     test('Get with module', () {
-      expect(Modular.get<AppBloc>(module: AppModule), isA<AppBloc>());
-      expect(Modular.get<HomeBloc>(module: HomeModule), isA<HomeBloc>());
+      expect(Modular.get<AppBloc>(module: 'AppModule'), isA<AppBloc>());
+      expect(Modular.get<HomeBloc>(module: 'HomeModule'), isA<HomeBloc>());
     });
 
     test('Inject not found with module', () {
       expect(() {
-        Modular.get<HomeBloc>(module: AppModule);
+        Modular.get<HomeBloc>(module: 'AppModule');
       }, throwsA(isA<ModularError>()));
       expect(() {
-        Modular.get<AppBloc>(module: HomeModule);
+        Modular.get<AppBloc>(module: 'HomeModule');
       }, throwsA(isA<ModularError>()));
     });
 
@@ -41,6 +42,12 @@ void main() {
       expect(() {
         Modular.get<HomeModule>();
       }, throwsA(isA<ModularError>()));
+    });
+
+    test('Get Interface', () {
+      expect(Modular.get<LocalStorageSharePreference>(),
+          isA<LocalStorageSharePreference>());
+      expect(Modular.get<ILocalStorage>(), isA<LocalStorageSharePreference>());
     });
   });
 }